test/requirements.txt
absl-py==2.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
importlib_metadata==8.0.0
isort==5.13.2
platformdirs==4.2.2
requests==2.32.3
smmap==5.0.1
tomli==2.0.1
urllib3==2.2.2
yapf==0.40.2
zipp==3.19.2
